Sparse input format.
The training data is now stored in sparse format.
The data reader converts to dense format so the neural
network computations work as usual.
Toolset functions related to unsupervised data removed, no need for them.
The input vector has 0's instead of -1's for positions not corresponding
to the word id.